+++ Sibusten.Philomena.Downloader.Cmd/Commands/DownloadCommand.cs
@@ -2,7 +2,12 @@ using System;
 using System.Collections.Generic;
 using System.CommandLine;
 using System.Threading.Tasks;
+using ShellProgressBar;
+using Sibusten.Philomena.Client;
+using Sibusten.Philomena.Client.Utilities;
 using Sibusten.Philomena.Downloader.Cmd.Commands.Arguments;
+using Sibusten.Philomena.Downloader.Cmd.Reporters;
+using Sibusten.Philomena.Downloader.Reporters;
 using Sibusten.Philomena.Downloader.Settings;
 
 namespace Sibusten.Philomena.Downloader.Cmd.Commands
@@ -61,8 +66,12 @@ namespace Sibusten.Philomena.Downloader.Cmd.Commands
             // Download images on each booru
             foreach (BooruConfig booruConfig in booruConfigs)
             {
+                IImageDownloadReporter reporter = new SingleLineImageDownloadConsoleReporter();
+
+                IProgress<ImageDownloadProgressInfo> progress = new SyncProgress<ImageDownloadProgressInfo>(reporter.ReportProgress);
+
                 ImageDownloader downloader = new ImageDownloader(booruConfig, searchConfig);
-                await downloader.StartDownload();
+                await downloader.StartDownload(progress: progress);
             }
         }
     }