+++ Sibusten.Philomena.Downloader/Reporters/IImageDownloadReporter.cs
@@ -1,0 +1,9 @@
+using Sibusten.Philomena.Client;
+
+namespace Sibusten.Philomena.Downloader.Reporters
+{
+    public interface IImageDownloadReporter
+    {
+        void ReportProgress(ImageDownloadProgressInfo progress);
+    }
+}